Alerts implemented to indicate errors, succeesses have framework for something to be done with them in place, but currently go silently.

+++ app/notes/notes.service.js
@@ -19,16 +19,48 @@
         note: note
       });
       notesPromise.then(function(res){
-        service.notes.unshift(res.data.note);
-        service.note = res.data.note;
+        if(res.status === 200) {
+          service.success();
+          service.notes.unshift(res.data.note);
+          service.note = res.data.note;
+        }
+        else{
+          service.error(res.data);
+        }
+
+      },function(res){
+        if(res.status === 200) {
+          service.success();
+          service.notes.unshift(res.data.note);
+          service.note = res.data.note;
+        }
+        else{
+          service.error(res.data|| 'Request Failed');
+        }
       });
       return notesPromise;
     };
     service.delete = function(note) {
       var notesPromise = $http.delete('https://meganote.herokuapp.com/notes/' + note._id, { note: note});
       notesPromise.then(function(res){
-        service.removeById(res.data._id);
-        service.notes = res.data;
+        if(res.status === 200) {
+          service.success();
+          service.removeById(res.data._id);
+        }
+        else{
+          service.error(res.data);
+        }
+
+
+      },function(res){
+        if(res.status === 200) {
+          service.success();
+          service.notes.unshift(res.data.note);
+          service.note = res.data.note;
+        }
+        else{
+          service.error(res.data|| 'Request Failed');
+        }
       });
       return notesPromise;
 
@@ -36,8 +68,23 @@
     service.update = function(note) {
       var notesPromise = $http.put('https://meganote.herokuapp.com/notes/' + note._id, { note: note});
       notesPromise.then(function(res){
-        service.removeById(res.data.note._id);
-        service.notes.unshift(res.data.note);
+        if(res.status === 200) {
+          service.success();
+          service.removeById(res.data.note._id);
+          service.notes.unshift(res.data.note);
+        }
+        else{
+          NotesService.error(res.data);
+        }
+      },function(res){
+        if(res.status === 200) {
+          service.success();
+          service.notes.unshift(res.data.note);
+          service.note = res.data.note;
+        }
+        else{
+          service.error(res.data || 'Request Failed');
+        }
       });
       return notesPromise;
 
@@ -48,6 +95,12 @@
           return service.notes.splice(i,1);
         }
       }
+    };
+    service.error = function(errorData) {
+      alert(errorData);
+    };
+    service.success = function(){
+
     };
   }
 })();
